+++ src/geometry.rs
