+++ src/util.rs
@@ -1,8 +1,10 @@
+use chrono::{TimeZone, UTC};
 use futures::{Future, Poll, Sink, Stream};
 use futures::stream::{self, Then};
 use futures::sync::mpsc::{self, Receiver};
 use hyper;
 use oauthcli::{OAuthAuthorizationHeader, ParseOAuthAuthorizationHeaderError};
+use serde::de::{Deserialize, Deserializer, Error};
 use std::fmt;
 use std::io::{self, BufRead};
 use std::str::FromStr;
@@ -10,6 +12,7 @@ use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::thread;
 use std::time::{Duration, Instant};
+use types::DateTime;
 
 /// A stream over each line on a `BufRead`.
 pub type Lines = Then<
@@ -28,6 +31,100 @@ pub struct Timeout {
     is_active: Arc<AtomicBool>,
 }
 
+macro_rules! string_enums {
+    (
+        $(
+            $(#[$attr:meta])*
+            pub enum $E:ident {
+                $(
+                    $(#[$v_attr:meta])*
+                    :$V:ident($by:expr) // The leading (ugly) colon is to suppress local ambiguity error.
+                ),*;
+                $(#[$u_attr:meta])*
+                :$U:ident(_),
+            }
+        )*
+    ) => {
+        $(
+            $(#[$attr])*
+            pub enum $E {
+                $(
+                    $(#[$v_attr])*
+                    $V,
+                )*
+                $(#[$u_attr])*
+                $U(String),
+            }
+
+            impl ::serde::Deserialize for $E {
+                fn deserialize<D: ::serde::Deserializer>(d: D) -> ::std::result::Result<Self, D::Error> {
+                    struct V;
+
+                    impl ::serde::de::Visitor for V {
+                        type Value = $E;
+
+                        fn visit_str<E>(self, s: &str) -> ::std::result::Result<$E, E> {
+                            match s {
+                                $($by => Ok($E::$V),)*
+                                _ => Ok($E::$U(s.to_owned())),
+                            }
+                        }
+
+                        fn visit_string<E>(self, s: String) -> ::std::result::Result<$E, E> {
+                            match s.as_str() {
+                                $($by => Ok($E::$V),)*
+                                _ => Ok($E::$U(s)),
+                            }
+                        }
+
+                        fn expecting(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
+                            write!(f, "a string")
+                        }
+                    }
+
+                    d.deserialize_string(V)
+                }
+            }
+
+            impl ::std::convert::AsRef<str> for $E {
+                fn as_ref(&self) -> &str {
+                    match *self {
+                        $($E::$V => $by,)*
+                        $E::$U(ref s) => s,
+                    }
+                }
+            }
+
+            impl ::std::cmp::PartialEq for $E {
+                fn eq(&self, other: &$E) -> bool {
+                    match *self {
+                        $($E::$V => match *other {
+                            $E::$V => true,
+                            $E::$U(ref s) if $by == s => true,
+                            _ => false,
+                        },)*
+                        $E::$U(ref s) => match *other {
+                            $($E::$V => $by == s,)*
+                            $E::$U(ref t) => s == t,
+                        },
+                    }
+                }
+            }
+
+            impl ::std::hash::Hash for $E {
+                fn hash<H: ::std::hash::Hasher>(&self, state: &mut H) {
+                    match *self {
+                        $($E::$V => $by.hash(state),)*
+                        $E::$U(ref s) => s.hash(state),
+                    }
+                }
+            }
+
+            impl ::std::cmp::Eq for $E {}
+        )*
+    }
+}
+
 /// Returns a stream over each line on `a`.
 #[allow(unused_variables)]
 pub fn lines<A: BufRead + Send + 'static>(a: A) -> Lines {
@@ -44,6 +141,14 @@ pub fn lines<A: BufRead + Send + 'static>(a: A) -> Lines {
     }
 
     rx.then(thener as _)
+}
+
+pub fn parse_datetime(s: &str) -> ::chrono::format::ParseResult<DateTime> {
+    UTC.datetime_from_str(s, "%a %b %e %H:%M:%S %z %Y")
+}
+
+pub fn deserialize_datetime<D: Deserializer>(d: D) -> Result<DateTime, D::Error> {
+    parse_datetime(&String::deserialize(d)?).map_err(|e| D::Error::custom(e.to_string()))
 }
 
 impl FromStr for OAuthHeaderWrapper {
